Restoring the Timeline
A massive chronokinetic fracture has scrambled all the events in human history in a random order.
Sadly, it may be possible that the exact original timeline may be irretrievable. However, as long as the overall trajectory of human society remains intact, this may not lead to unconscionable repercussions.
The original timeline can be represented as a sequence of positive integers A_{1}, A_{2},..., A_{n_{1}}
Post-fracture, the new timeline is a new sequence of events B_{1}, B_{2},..., B_{n_{2}}.
We have a time machine that we can use to send us to the very beginning of the timeline, having us right before event B_{1}. After being sent back, we can only move forward, not backward. This allows us to skip huge chunks of time to more efficiently fix the timeline, but we cannot backtrack if we make any errors in the process.
Initially, we are looking at B1. We have a few actions we can take:
- We can manually cause our own custom event in the timeline with an integer value of k. This will cost 2k energy from our time machine. This will place us after our manufactured event in the timeline.
- We can prevent the next event in the timeline from happening, costing k energy from our time machine. This will place us right before the next event in the timeline after the event we just deleted.
- We can prevent the next event from happening, altering it into a new event with value k. This costs 2|k-B_{i}| energy, where B_{i} is the value of the event we are preventing (this is assuming we have traveled to right before event B_{i} occurs). This places us after the newly manufactured event in the timeline.
- We can perform ultimate timeline shenanigans and simply swap the following two events that will occur in the timeline, costing our time machine |B_{i} - B_{i} + 1 | where B_{i} is the next event in the timeline and B_{i+1} is the event immediately after it. This places us immediately before event B_{i+2} or immediately after the two swapped events.
Given the original and the fractured sequences, what is the minimal energy needed to convert the fractured sequence into the original sequence?

Input Format
Input begins with a line containing two space-separated integers, N_{1} and N_{2}, indicating the lengths of the original sequence and the fractured sequence, respectively.
N_{1} lines follow, each containing a single integer A_{i}, indicating the integer value of the ith event in the original sequence.
N_{2} lines follow, each containing a single integer B_{j} indicating the integer value of the jth event in the fractured sequence.

Constraints
1 <= N_{1}, N_{2} <= 5.5 * 10 ^ 3
1 <= A_{i}, A_{j} <= 10 ^ 9 for all 1 <= i <= N_{1}, 1 <= j <= N_{2}

Output Format
Output a line containing a single integer indicating the minimum energy needed to restore the original timeline.

Sample Input 0
3 3
6
10
4
2
11
12

Sample Output 0
24

Explanation 0
1.) PREVENT event with value 2 for 2 energy. Current Timeline: [11], 12
2.) ALTER event 11 into event with value 6 for 10 energy. Current Timeline: 6, [12]
3.) ALTER event 12 into event with value 10 for 4 energy. Current Timeline: 6, 10, []
4.) CAUSE event with value 4 for 8 energy. Current Timeline: 6, 10, 4

Sample Input 1
5 5
5
17
7
7
9
18
11
3
9
7

Sample Output 1
25

Explanation 1
1.) CAUSE event with value 5 for 10 energy. Current Timeline: 5, [18], 11, 3, 9, 7
2.) ALTER event 18 into event with value 17 for 2 energy. Current Timeline: 5, 17, [11], 3, 9, 7
3.) ALTER event 11 into event with value 7 for 8 energy. Current Timeline: 5, 17, 7, [3], 9, 7
4.) PREVENT event with value 3 for 3 energy. Current Timeline: 5, 17, 7, [9], 7
5.) SWAP events 9 and 7 for 9 energy. Current Timeline: 5, 17, 7, 7, 9